% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assay_abbr}
\alias{assay_abbr}
\title{assay abbreviations}
\format{
An object of class \code{character} of length 19.
}
\usage{
assay_abbr
}
\description{
assay abbreviations
}
\keyword{datasets}
