% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assay_codes}
\alias{assay_codes}
\title{Motrpac assay codes}
\format{
A data frame with assay codes
\describe{
\item{omics_code}{Main OMIC group}
\item{submission_code}{Assay submission code}
\item{assay}{Assay release code}
\item{assay_code}{Assay release code, same as assay}
\item{assay_name}{Assay long name}
\item{cas_code}{MoTrPAC chemical analysis site}
\item{assay_abbreviation}{Assay abbreviation}
\item{assay_hex_colour}{Assay hex colour}
\item{assay_short_text}{Abbreviated name specific to the assay that can be used in graphs and tables as needed}
\item{omics_text}{OMIC class measured by the given assay}
\item{ome_text}{OME category measured by the given assay}
}
}
\usage{
assay_codes
}
\description{
Motrpac assay codes
}
\examples{
\dontrun{
 assay_codes
}
}
\keyword{datasets}
