% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bic_animal_tissue_code}
\alias{bic_animal_tissue_code}
\title{BIC Tissue Code and Name}
\format{data frame}
\usage{
bic_animal_tissue_code
}
\description{
BIC Tissue Code and Name
}
\keyword{datasets}
