% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_qc.R
\name{check_crossfile_olink_validation}
\alias{check_crossfile_olink_validation}
\title{Cross-validate OLINK files}
\usage{
check_crossfile_olink_validation(
  r_o,
  m_s,
  m_p,
  return_n_issues = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{r_o}{(data.frame) results df}

\item{m_s}{(char) metadata_sample df}

\item{m_p}{(char) metadata_proteins df}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
Check values from results are vailable in both metdata proteins
and metadata samples files
}
\examples{
{
\dontrun{
check_crossfile_olink_validation(r_o = results, m_s = metadata_samples, m_p = metadata_proteins)
}
}
}
