% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{check_crossfile_validation}
\alias{check_crossfile_validation}
\title{Cross-validate assay files (Olink and LAB)}
\usage{
check_crossfile_validation(
  r_o,
  m_s,
  m_p,
  assay_type = c("olink", "lab"),
  return_n_issues = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{r_o}{(data.frame) Results data frame.}

\item{m_s}{(data.frame) Metadata sample data frame.}

\item{m_p}{(data.frame) Metadata analyte/protein data frame.}

\item{assay_type}{(character) The type of assay, either \code{"olink"} or \code{"lab"}.}

\item{return_n_issues}{(logical) If \code{TRUE}, returns the number of issues.}

\item{verbose}{(logical) If \code{TRUE} (default), displays messages during the checking process.}
}
\value{
(int) Number of issues identified if \code{return_n_issues} is \code{TRUE}.
}
\description{
Checks that values from the results file are available in both the
metadata analyte/protein and metadata sample files.
}
\examples{
{
\dontrun{
# For Olink data
check_crossfile_validation(r_o = results_olink,
                           m_s = metadata_samples_olink,
                           m_p = metadata_proteins_olink,
                           assay_type = "olink")

# For LAB data
check_crossfile_validation(r_o = results_lab,
                           m_s = metadata_samples_lab,
                           m_p = metadata_analyte_lab,
                           assay_type = "lab")
}
}
}
