% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{check_failedsamples}
\alias{check_failedsamples}
\title{check failed samples file for reported missing vial label ids}
\usage{
check_failedsamples(input_results_folder, verbose = TRUE)
}
\arguments{
\item{input_results_folder}{(char) input path folder}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(vector) failed reported ids
}
\description{
check failed samples file for reported missing vial label ids
}
