% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imm_qc.R
\name{check_metadata_analyte}
\alias{check_metadata_analyte}
\title{Check analyte metadata file}
\usage{
check_metadata_analyte(
  df,
  return_n_issues = FALSE,
  validate_uniprot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{(data.frame) The metadata_analyte data frame to check.}

\item{return_n_issues}{(logical) If \code{TRUE}, returns the number of issues found.}

\item{validate_uniprot}{(logical) If \code{TRUE}, checks if all Uniprot IDs are valid by connecting to the Uniprot database. Note: This may take several minutes depending on the number of IDs.}

\item{verbose}{(logical) If \code{TRUE} (default), displays messages during the checking process.}
}
\value{
(int) The number of issues identified if \code{return_n_issues} is \code{TRUE}.
}
\description{
Checks whether the metadata_analyte file follows the required guidelines.
}
\examples{
\dontrun{
check_metadata_analyte(df = metadata_analyte)
}
}
