% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{check_metadata_phase_file}
\alias{check_metadata_phase_file}
\title{check metadata phase file}
\usage{
check_metadata_phase_file(input_results_folder, verbose)
}
\arguments{
\item{input_results_folder}{(char) input path folder}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(logical) \code{TRUE} exist, \code{FALSE} does not
}
\description{
check the existence of the metadata phase file
}
