% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_qc.R
\name{check_metadata_proteins}
\alias{check_metadata_proteins}
\title{Check olink metadata proteins file}
\usage{
check_metadata_proteins(
  df,
  return_n_issues = FALSE,
  validate_uniprot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{(data.frame) metadata_proteins df}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues.}

\item{validate_uniprot}{(logical) if \code{TRUE}, check if all uniprot ids are valid
connecting to Uniprot. Note: depending on the number of ids, it might take
several finish to complete the validation}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether metadata-proteins.txt file is following guidelines
}
\examples{
{
\dontrun{
check_metadata_proteins(df = metadata_proteins)
}
}
}
