% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imm_qc.R
\name{check_metadata_samples_imm}
\alias{check_metadata_samples_imm}
\title{Check IMM metadata samples file}
\usage{
check_metadata_samples_imm(df, return_n_issues = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{(data.frame) IMM metadata samples data}

\item{return_n_issues}{(logical) If \code{TRUE}, returns the number of issues identified.}

\item{verbose}{(logical) If \code{TRUE} (default), displays messages during the checking process.}
}
\value{
(int) Number of issues identified if \code{return_n_issues} is \code{TRUE}.
}
\description{
Checks whether the metadata_sample file for IMM assays follows the required guidelines.
}
\examples{
\dontrun{
check_metadata_samples_imm(df = metadata_sample_named_CK_plasma)
}
}
