% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_qc.R
\name{check_metadata_samples_olink}
\alias{check_metadata_samples_olink}
\title{check olink metadata samples file}
\usage{
check_metadata_samples_olink(df, return_n_issues = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{(data.frame) olink metadata samples ata}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues.}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether metadata_sample is following guidelines
}
\examples{
{
\dontrun{
check_metadata_samples(df = metadata_sample_named)
}
}
}
