% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Check for Missing Values in a Data Frame Column}
\usage{
check_missing_values(df, column)
}
\arguments{
\item{df}{A data frame in which the column to be checked is located.}

\item{column}{The name of the column to check for missing values, as a string.}
}
\value{
A boolean value; TRUE if the specified column has no missing values, FALSE if it does.
}
\description{
This function checks for missing values in a specified column of a data frame.
It returns TRUE if there are no missing values in the column, and FALSE otherwise.
}
\examples{
data <- data.frame(a = c(1, 2, NA, 4), b = c("A", "B", "C", "D"))
check_missing_values(data, "a") # returns TRUE
check_missing_values(data, "b") # returns FALSE

}
