% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_qc.R
\name{check_ratio_proteomics}
\alias{check_ratio_proteomics}
\title{check proteomics ratio file}
\usage{
check_ratio_proteomics(
  df_ratio,
  isPTM,
  f_proof = TRUE,
  output_prefix = "ratio-file",
  out_qc_folder = NULL,
  printPDF = TRUE,
  return_n_issues = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df_ratio}{(data.frame) proteomics ratio results data frame (required)}

\item{isPTM}{(logical) \code{TRUE} if a PTM results file}

\item{f_proof}{(logical) \code{TRUE} (default) to print out distribution and NA plots}

\item{output_prefix}{(char) if \code{f_proof = TRUE}, provide a prefix for the output name}

\item{out_qc_folder}{(char) if \verb{f_proof is TRUE}, a folder path can be provided
(otherwise print in current working directory)}

\item{printPDF}{(logical) if \code{TRUE} (default print plots to pdf)}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues.}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether the proteomics ratio results files is following guidelines
}
\examples{
{
test <- check_ratio_proteomics(df_ratio = metadata_metabolites_named,
isPTM =  TRUE, return_n_issues = TRUE, verbose = FALSE)
# "test" should be NULL
}
}
