% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{check_results}
\alias{check_results}
\title{check results}
\usage{
check_results(r_m, m_s, m_m, verbose = TRUE)
}
\arguments{
\item{r_m}{(data.frame) results df}

\item{m_s}{(char) metadata_sample df}

\item{m_m}{(char) metadata_metabolites df}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether results file is following guidelines
}
