% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_qc.R
\name{check_results_assays}
\alias{check_results_assays}
\title{Check results file for assays}
\usage{
check_results_assays(
  df,
  assay_type = c("olink", "imm"),
  return_n_issues = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{(data.frame) The results data frame to check.}

\item{assay_type}{(character) The type of assay, either \code{"olink"} or \code{"imm"}.}

\item{return_n_issues}{(logical) If \code{TRUE}, returns the number of issues identified.}

\item{verbose}{(logical) If \code{TRUE} (default), displays messages during the checking process.}
}
\value{
(int) Number of issues identified if \code{return_n_issues} is \code{TRUE}.
}
\description{
Checks whether the results file follows guidelines for Olink and IMM assays.
}
\examples{
\dontrun{
check_results_assays(df = results_df, assay_type = "olink")
check_results_assays(df = results_df, assay_type = "imm")
}
}
