% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_qc.R
\name{check_results_olink}
\alias{check_results_olink}
\title{Check olink results}
\usage{
check_results_olink(df, return_n_issues = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{(data.frame) olink results data frame}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues.}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether olink results file follows guidelines
}
\examples{
{
\dontrun{
check_metadata_samples(df = olink_results)
}
}
}
