% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_qc.R
\name{check_rii_proteomics}
\alias{check_rii_proteomics}
\title{check proteomics reported ion intensity file}
\usage{
check_rii_proteomics(
  df_rri,
  isPTM,
  f_proof = TRUE,
  output_prefix = "rii-file",
  out_qc_folder = NULL,
  return_n_issues = FALSE,
  printPDF = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{df_rri}{(data.frame) proteomics rri data frame (required)}

\item{isPTM}{(logical) \code{TRUE} if a PTM results file}

\item{f_proof}{(logical) \code{TRUE} (default) to print out distribution and NA plots}

\item{output_prefix}{(char) if \code{f_proof = TRUE}, provide a prefix for the output name}

\item{out_qc_folder}{(char) if \verb{f_proof is TRUE}, a folder path can be provided
(otherwise print in current working directory)}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues.}

\item{printPDF}{(logical) if \code{TRUE} (default print plots to pdf)}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether the proteomics rri results files is following guidelines
}
\examples{
{
test <- check_rii_proteomics(df_rri = metadata_metabolites_named,
isPTM =  TRUE, return_n_issues = TRUE, verbose = FALSE)
# "test" should be NULL
}
}
