% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_qc.R
\name{check_vial_metadata_proteomics}
\alias{check_vial_metadata_proteomics}
\title{check proteomics vial metadata file}
\usage{
check_vial_metadata_proteomics(df_vm, return_n_issues = FALSE, verbose = TRUE)
}
\arguments{
\item{df_vm}{(data.frame) proteomics vial_label data frame (required)}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues.}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether the proteomics rri results files is following guidelines
}
\examples{
{
test <- check_vial_metadata_proteomics(df_vm = metadata_metabolites_named,
return_n_issues = TRUE, verbose = FALSE)
# "test" should be NULL
}
}
