% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{check_viallabel_dmaqc}
\alias{check_viallabel_dmaqc}
\title{validate vial label from DMAQC}
\usage{
check_viallabel_dmaqc(
  vl_submitted,
  dmaqc_shipping_info,
  tissue_code,
  cas,
  phase,
  failed_samples,
  verbose = TRUE
)
}
\arguments{
\item{vl_submitted}{(vector) results df}

\item{dmaqc_shipping_info}{(data.frame) dmaqc shipping information}

\item{tissue_code}{(char) tissue code}

\item{cas}{(char) CAS code}

\item{phase}{(char) phase code}

\item{failed_samples}{(char) metadata_metabolites df}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
validate vial label from DMAQC
}
