% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_folder}
\alias{create_folder}
\title{Create folder}
\usage{
create_folder(folder_name = NULL, verbose = FALSE)
}
\arguments{
\item{folder_name}{(chr) folder name}

\item{verbose}{(logical) \code{TRUE} shows messages (default \code{FALSE})}
}
\description{
Create a directory if it doesn't exist. If no argument is provided,
it returns the current working directory
}
\examples{
{
create_folder(folder_name = NULL)
# Or use this one for a real folder:
# create_folder(folder_name = "testing")
}
}
