% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_required_columns}
\alias{filter_required_columns}
\title{filter required columns only}
\usage{
filter_required_columns(
  df,
  type = c("m_m", "m_s", "v_m", "olproteins", "olsamples", "immanalytes", "immsamples"),
  name_id = NULL,
  verbose = TRUE
)
}
\arguments{
\item{df}{(data.frame) metadata_metabolites}

\item{type}{(char) Type of file to filter columns:
\itemize{
\item \code{m_m}: metadata metabolites
\item \code{m_s}: metadata samples
\item \code{v_m}: proteomics vial_metadata
\item \code{olproteins}: olink metadata proteins
\item \code{olsamples}: olink metadata samples
}}

\item{name_id}{(char) specify whether \code{named} or \code{unnamed} files}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(data.frame) filtered data frame with only the required columns
}
\description{
it returns a data frame with only the required columns for metabolomics and proteomics
}
\examples{
{
df_filtered <- filter_required_columns(df = metadata_metabolites_named, name_id = "named")
}
}
