% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_full_path2batch}
\alias{get_full_path2batch}
\title{Get full path to the batch folder}
\usage{
get_full_path2batch(input_results_folder)
}
\arguments{
\item{input_results_folder}{(char) path to the PROCESSED/RESULTS folder to check}
}
\value{
(char) Full path to the \code{BATCH#_YYYYMMDD} folder
}
\description{
Get the full path to the batch folder
}
