% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_abbr}
\alias{group_abbr}
\title{group abbreviations}
\format{
An object of class \code{character} of length 6.
}
\usage{
group_abbr
}
\description{
group abbreviations
}
\keyword{datasets}
