% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_qc.R
\name{load_lab_batch}
\alias{load_lab_batch}
\title{Load and Process LAB Batch Data}
\usage{
load_lab_batch(input_results_folder, verbose = TRUE)
}
\arguments{
\item{input_results_folder}{A string representing the path to the folder
containing LAB batch data to be loaded and processed.}

\item{verbose}{Logical; if \code{TRUE}, prints detailed messages
during the loading process.}
}
\value{
A list containing data frames for metadata of analytes (\code{m_a}),
metadata of samples (\code{m_s}), and LAB results (\code{r_o}). If certain files
are not available, the corresponding entries in the list will be \code{NULL}.
}
\description{
This function loads LAB batch data from the specified input directory.
It performs quality checks on the data and loads specific files related to
LAB data, including metadata for analytes and samples, and the results file.
It also integrates validation checks and warns if there are too many issues
identified in the data.
}
\examples{
\dontrun{
list_of_df <- load_lab_batch(input_results_folder = "/path/to/PROCESSED_YYYYMMDD/")
}
}
