% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_qc.R
\name{load_olink_batch}
\alias{load_olink_batch}
\title{Load and Process Olink Batch Data}
\usage{
load_olink_batch(input_results_folder, verbose = TRUE)
}
\arguments{
\item{input_results_folder}{A string representing the path to the folder
containing Olink batch data to be loaded and processed.}

\item{verbose}{Logical; if \code{TRUE}, prints detailed messages
during the loading process.}
}
\value{
A list containing data frames for metadata of proteins (m_p),
metadata of samples (m_s), and Olink results (r_o). If certain files
are not available, the corresponding entries in the list will be NULL.
}
\description{
This function loads Olink batch data from the specified input directory.
It performs quality checks on the data and loads specific files related to
Olink data, including metadata for proteins and samples, and the results file.
It also integrates validation checks and warns if there are too many issues
identified in the data.
}
\examples{
\dontrun{
list_of_df <- load_olink_batch(input_results_folder = "/path/to/PROCESSED_YYYYMMDD/")
}
}
