% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_qc.R
\name{load_proteomics}
\alias{load_proteomics}
\title{Load Proteomics batch}
\usage{
load_proteomics(input_results_folder, isPTM, verbose = TRUE)
}
\arguments{
\item{input_results_folder}{(char) path to the PROCESSED folder to check}

\item{isPTM}{(logical) \code{TRUE} if it is Post-Translational-Modification proteomics assay}

\item{verbose}{(logical) \code{TRUE} (default) prints QC details.}
}
\value{
(List) of data frames: rii,
}
\description{
Load a proteomics batch
}
