% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{merge_all_metabolomics}
\alias{merge_all_metabolomics}
\title{Merge all metabolomics files}
\usage{
merge_all_metabolomics(m_m_n, m_m_u = NULL, m_s_n, r_n, r_u = NULL, phase)
}
\arguments{
\item{m_m_n}{(metabolomics metadata named)}

\item{m_m_u}{(metabolomics metadata unnamed)}

\item{m_s_n}{(metabolomics sample named)}

\item{r_n}{(results named)}

\item{r_u}{(results unnamed)}

\item{phase}{(MoTrPAC Animal phase. Eg. PASS1A-06)}
}
\value{
(data.frame) Merged data frame long format
}
\description{
Merge all metabolomics datasets, including "results" and
"metadata" files, for both targeted and untargeted datasets
}
\examples{
plasma.untargeted.merged <- merge_all_metabolomics(
       m_m_n = metadata_metabolites_named,
       m_m_u = metadata_metabolites_unnamed,
       m_s_n = metadata_sample_named,
       r_n = results_named,
       r_u = results_unnamed,
       phase = "PASS1A-06")
}
