% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{merge_metabolomics_metadata}
\alias{merge_metabolomics_metadata}
\title{Merge metabolomics metadata named and unnamed}
\usage{
merge_metabolomics_metadata(m_m_n, m_m_u)
}
\arguments{
\item{m_m_n}{(df) metabolomics metadata named}

\item{m_m_u}{(char) metabolomics metadata unnamed}
}
\value{
(data.frame) merged metadata metabolites
}
\description{
Merge metabolomics metadata
}
\examples{
{
m_m <- merge_metabolomics_metadata(m_m_n = metadata_metabolites_named,
                                   m_m_u = metadata_metabolites_unnamed)
}
}
