% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{merge_phenotype_metabolomics}
\alias{merge_phenotype_metabolomics}
\title{Merge phenotypic and metabolics results}
\usage{
merge_phenotype_metabolomics(df_long)
}
\arguments{
\item{df_long}{(data.frame) Long format of a metabolomics merged results}
}
\value{
(data.frame) Merged file, including the following columns:
\describe{
\item{sample_id}{Sample Id, including vial_label and site specific QC ids}
\item{sample_type}{Metabolomics sample types. Check metabolomics data transfer guidelines}
\item{sample_order}{Order of injection on Mass Spec}
\item{metabolite_name}{Given name by every lab}
\item{refmet_name}{Map of the metabolite name to the Metabolomics RefMet database}
\item{mz}{mass over charge}
\item{rt}{retention time}
\item{formula}{chemical formula}
\item{neutral_mass}{neutral mass}
\item{id}{type of metabolite identification: "named", "unnamed"}
\item{metabolite}{Merge "refmet" for "named" metabolites and "metabolite_name" for "unnamed" metabolites}
\item{quantification}{Untargeted: Peak area, Targeted: absolute concentration (check "experimentalDetails" for unit)}
\item{tissue_code}{MoTrPAC tissue code}
\item{tissue_name}{Tissue name}
\item{group_time_point}{Intervention group (Exercise / Control) +  time point}
\item{sex}{Animal Sex}
\item{site_code}{Chemical Analysis Site (CAS) short abbreviation}
\item{group}{Intervention group: Exercise / Control}
\item{condition}{Sex + group + time-point}
\item{bioreplicate}{Sex + group + time-point + sample_order}
}
}
\description{
Merge phenotypic data (phenotypes_pass1a06_short) and
metabolomics merged results and metadata
}
