% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabolomics_data_dictionary}
\alias{metabolomics_data_dictionary}
\title{Metabolomics data dictionary}
\format{
A data frame with > 2000 rows and 16 15 columns (variables):
\describe{
\item{CURRENT_REFMET_NAME}{Updated version of RefMet (after December 2020)}
\item{refmet_name}{Old version of refmet_name}
\item{metabolite_name}{Metabolite name provided by each laboratory}
\item{is_standard}{is a reference standard? 1: yes, 0: no}
\item{super_class}{Metabolite super class}
\item{main_class}{Metabolite main class}
\item{sub_class}{Metabolite sub class}
\item{formula}{Metabolite formula}
\item{exactmass}{Metaboliteexact mass}
\item{pubchem_cid}{Metabolite pubchem id}
\item{kegg_id}{Metabolite kegg id}
\item{inchi_key}{Metabolite inchi key}
\item{lm_id}{Metabolite lm id}
\item{hmdb_id}{Metabolite hmdb id}
\item{chebi_id}{Metabolite chebi id}
}
}
\usage{
metabolomics_data_dictionary
}
\description{
Metabolomics data dictionary
}
\examples{
\dontrun{
 metabolomics_data_dictionary
}
}
\keyword{datasets}
