% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_metabolites_named}
\alias{metadata_metabolites_named}
\title{Metadata Metabolites test dataset for NAMED metabolites}
\format{
A data frame with 53 rows and 6 columns:
\describe{
\item{metabolite_name}{metabolite names}
\item{refmet_name}{RefMet id (metabolomics workbench)}
\item{mz}{mass over charge}
\item{rt}{retention time}
\item{formula}{chemical formula}
\item{neutral_mass}{neutral mass}
}
}
\usage{
metadata_metabolites_named
}
\description{
Metadata Metabolites test dataset for NAMED metabolites
}
\examples{
\dontrun{
 metadata_metabolites_named
}
}
\keyword{datasets}
