% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_sample_named}
\alias{metadata_sample_named}
\title{Metadata Sample test dataset for NAMED metabolites}
\format{A data frame with 97 rows and 4 columns:
\describe{
\item{sample_id}{metabolite names based on mz_rt}
\item{sample_type}{mass over charge}
\item{sample_order}{retention time}
\item{raw_file}{neutral mass}
}}
\usage{
metadata_sample_named
}
\description{
Metadata Sample test dataset for NAMED metabolites
}
\keyword{datasets}
