% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{open_file}
\alias{open_file}
\title{open files}
\usage{
open_file(input_results_folder, filepattern, verbose = TRUE)
}
\arguments{
\item{input_results_folder}{(char) input path folder}

\item{filepattern}{(char) regular expression to find a file in the file system
provided}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(list) list with data frame and flag about the status
}
\description{
open files and check that they are right
}
