% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotypes_pass1a06_short}
\alias{phenotypes_pass1a06_short}
\title{Motrpac Phenotypes PASS1A 6 Months}
\format{
A data frame with 8616 rows and 6 columns:
\describe{
\item{tissue_code}{MoTrPAC Tissue code}
\item{vial_label}{Vial label id (only available for MoTrPAC Samples)}
\item{tissue_name}{Tissue name}
\item{group_time_point}{Group time point}
\item{sex}{Animal sex}
\item{site_code}{cas}
}
}
\usage{
phenotypes_pass1a06_short
}
\description{
Motrpac Phenotypes PASS1A 6 Months
}
\examples{
\dontrun{
 phenotypes_pass1a06_short
}
}
\keyword{datasets}
