% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{plot_basic_metabolomics_qc}
\alias{plot_basic_metabolomics_qc}
\title{Plot Basic Metabolomics QC charts}
\usage{
plot_basic_metabolomics_qc(
  results,
  results_long,
  metametab = NULL,
  out_qc_folder = NULL,
  output_prefix,
  printPDF = TRUE,
  untargeted = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{results}{(df) metabolomics results (both named and unnamed already merged, if untargeted)}

\item{results_long}{(df) metabolomics results (both named and unnamed already merged, if untargeted, in long format)}

\item{metametab}{(df) metadata samples metadata (only untargeted sites)}

\item{out_qc_folder}{(char) output qc folder (it creates the folder if it doesn't exist)}

\item{output_prefix}{(char) prefix for the file name output (pdf file)}

\item{printPDF}{(logical) \code{TRUE} (default) prints pdf file}

\item{untargeted}{(logical) \code{TRUE} if the dataset is untargeted (named + unnamed metabolites)}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\description{
Plot intensity distributions, number of unique ids, NA values
and RT/MZ density maps (only untargeted)
}
