% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{proteomics_plots_rii}
\alias{proteomics_plots_rii}
\title{QC Plot of proteomics reporter ion intensity data}
\usage{
proteomics_plots_rii(
  all_vial_labels,
  all_samples,
  peprii,
  isPTM,
  v_m,
  out_qc_folder = NULL,
  output_prefix,
  printPDF,
  verbose
)
}
\arguments{
\item{all_vial_labels}{(vector) only vial_labels}

\item{all_samples}{(vector) all sample ids}

\item{peprii}{(char) Reporter Ion Intensity df}

\item{isPTM}{(logical) Is a PTM assay}

\item{v_m}{(df) sample metadata}

\item{out_qc_folder}{(char) if \verb{f_proof is TRUE}, a folder path can be provided
(otherwise print in current working directory)}

\item{output_prefix}{(char) provide a prefix for the output name}

\item{printPDF}{(logical) if \code{TRUE} (default print plots to pdf)}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether results file is following guidelines
}
\examples{
{
check_results(r_m = results_named, m_s = metadata_sample_named, m_m = metadata_metabolites_named)
}
}
