% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remove_empty_columns}
\alias{remove_empty_columns}
\title{remove empty columns}
\usage{
remove_empty_columns(df, verbose = TRUE)
}
\arguments{
\item{df}{(char) data frame}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(df) df without empty columns
}
\description{
remove empty columns
}
