% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{remove empty rows in data frame}
\usage{
remove_empty_rows(df, verbose = TRUE)
}
\arguments{
\item{df}{(char) data frame}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(df) df without empty columns
}
\description{
remove empty rows in data frame
}
