% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results_named}
\alias{results_named}
\title{Results test dataset for NAMED metabolites}
\format{
A data frame with 58 rows and 98 columns:
\describe{
\item{metabolite_name}{metabolite names}
\item{90001013104}{Sample id. Unit: Peak Area}
\item{90005013104}{sample id. Unit: Peak Area}
\item{90007013104}{sample id. Unit: Peak Area}
\item{90008013104}{sample id. Unit: Peak Area}
\item{90009013104}{sample id. Unit: Peak Area}
\item{90010013104}{sample id. Unit: Peak Area}
\item{90011013104}{sample id. Unit: Peak Area}
\item{90012013104}{sample id. Unit: Peak Area}
\item{90013013104}{sample id. Unit: Peak Area}
\item{90014013104}{sample id. Unit: Peak Area}
\item{90015013104}{sample id. Unit: Peak Area}
\item{90017013104}{sample id. Unit: Peak Area}
\item{90018013104}{sample id. Unit: Peak Area}
\item{90020013104}{sample id. Unit: Peak Area}
\item{90023013104}{sample id. Unit: Peak Area}
\item{90025013104}{sample id. Unit: Peak Area}
\item{90026013104}{sample id. Unit: Peak Area}
\item{90027013104}{sample id. Unit: Peak Area}
\item{90028013104}{sample id. Unit: Peak Area}
\item{90029013104}{sample id. Unit: Peak Area}
\item{90031013104}{sample id. Unit: Peak Area}
\item{90032013104}{sample id. Unit: Peak Area}
\item{90033013104}{sample id. Unit: Peak Area}
\item{90034013104}{sample id. Unit: Peak Area}
\item{90037013104}{sample id. Unit: Peak Area}
\item{90038013104}{sample id. Unit: Peak Area}
\item{90039013104}{sample id. Unit: Peak Area}
\item{90040013104}{sample id. Unit: Peak Area}
\item{90042013104}{sample id. Unit: Peak Area}
\item{90043013104}{sample id. Unit: Peak Area}
\item{90044013104}{sample id. Unit: Peak Area}
\item{90045013104}{sample id. Unit: Peak Area}
\item{90046013104}{sample id. Unit: Peak Area}
\item{90047013104}{sample id. Unit: Peak Area}
\item{90048013104}{sample id. Unit: Peak Area}
\item{90050013104}{sample id. Unit: Peak Area}
\item{90052013104}{sample id. Unit: Peak Area}
\item{90053013104}{sample id. Unit: Peak Area}
\item{90109013104}{sample id. Unit: Peak Area}
\item{90110013104}{sample id. Unit: Peak Area}
\item{90112013104}{sample id. Unit: Peak Area}
\item{90114013104}{sample id. Unit: Peak Area}
\item{90115013104}{sample id. Unit: Peak Area}
\item{90117013104}{sample id. Unit: Peak Area}
\item{90118013104}{sample id. Unit: Peak Area}
\item{90119013104}{sample id. Unit: Peak Area}
\item{90120013104}{sample id. Unit: Peak Area}
\item{90121013104}{sample id. Unit: Peak Area}
\item{90122013104}{sample id. Unit: Peak Area}
\item{90123013104}{sample id. Unit: Peak Area}
\item{90124013104}{sample id. Unit: Peak Area}
\item{90126013104}{sample id. Unit: Peak Area}
\item{90127013104}{sample id. Unit: Peak Area}
\item{90128013104}{sample id. Unit: Peak Area}
\item{90129013104}{sample id. Unit: Peak Area}
\item{90133013104}{sample id. Unit: Peak Area}
\item{90134013104}{sample id. Unit: Peak Area}
\item{90135013104}{sample id. Unit: Peak Area}
\item{90136013104}{sample id. Unit: Peak Area}
\item{90138013104}{sample id. Unit: Peak Area}
\item{90139013104}{sample id. Unit: Peak Area}
\item{90140013104}{sample id. Unit: Peak Area}
\item{90141013104}{sample id. Unit: Peak Area}
\item{90142013104}{sample id. Unit: Peak Area}
\item{90143013104}{sample id. Unit: Peak Area}
\item{90144013104}{sample id. Unit: Peak Area}
\item{90145013104}{sample id. Unit: Peak Area}
\item{90146013104}{sample id. Unit: Peak Area}
\item{90147013104}{sample id. Unit: Peak Area}
\item{90148013104}{sample id. Unit: Peak Area}
\item{90150013104}{sample id. Unit: Peak Area}
\item{90152013104}{sample id. Unit: Peak Area}
\item{90155013104}{sample id. Unit: Peak Area}
\item{90156013104}{sample id. Unit: Peak Area}
\item{90157013104}{sample id. Unit: Peak Area}
\item{90159013104}{sample id. Unit: Peak Area}
\item{90160013104}{sample id. Unit: Peak Area}
\item{CS00000MP-10}{sample id. Unit: Peak Area}
\item{CS00000MP-11}{sample id. Unit: Peak Area}
\item{CS00000MP-12}{sample id. Unit: Peak Area}
\item{CS00000MP-2}{sample id. Unit: Peak Area}
\item{CS00000MP-3}{sample id. Unit: Peak Area}
\item{CS00000MP-4}{sample id. Unit: Peak Area}
\item{CS00000MP-5}{sample id. Unit: Peak Area}
\item{CS00000MP-6}{sample id. Unit: Peak Area}
\item{CS00000MP-7}{sample id. Unit: Peak Area}
\item{CS00000MP-8}{sample id. Unit: Peak Area}
\item{CS00000MP-9}{sample id. Unit: Peak Area}
\item{CS000QCMP-1}{sample id. Unit: Peak Area}
\item{CS000QCMP-2}{sample id. Unit: Peak Area}
\item{CS000QCMP-4}{sample id. Unit: Peak Area}
\item{CS000QCMP-5}{sample id. Unit: Peak Area}
\item{CS000QCMP-6}{sample id. Unit: Peak Area}
\item{CSMR80014-1}{sample id. Unit: Peak Area}
\item{CSMR80014-2}{sample id. Unit: Peak Area}
\item{CSMR80015-1}{sample id. Unit: Peak Area}
\item{CSMR80015-2}{sample id. Unit: Peak Area}
}
}
\usage{
results_named
}
\description{
Results test dataset for NAMED metabolites
}
\examples{
\dontrun{
 results_named
}
}
\keyword{datasets}
