% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_abbr}
\alias{tissue_abbr}
\title{tissue abbreviations}
\format{
An object of class \code{character} of length 54.
}
\usage{
tissue_abbr
}
\description{
tissue abbreviations
}
\keyword{datasets}
