% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_assay}
\alias{validate_assay}
\title{extract ASSAY from input folder path}
\usage{
validate_assay(input_results_folder)
}
\arguments{
\item{input_results_folder}{(char) input_results_folder path}
}
\value{
(vector) ASSAY code
}
\description{
extract ASSAY from input folder path
}
