% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_qc.R
\name{validate_lab}
\alias{validate_lab}
\title{Validate LAB Assay Data}
\usage{
validate_lab(
  input_results_folder,
  cas,
  return_n_issues = FALSE,
  full_report = FALSE,
  f_proof = FALSE,
  printPDF = FALSE,
  out_qc_folder = NULL,
  dmaqc_shipping_info = NULL,
  dmaqc_phase2validate = FALSE,
  validate_uniprot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input_results_folder}{A string representing the path to the folder
containing LAB assay results to be validated.}

\item{cas}{A character string indicating the CAS number.}

\item{return_n_issues}{Logical; if \code{TRUE}, the function returns the number of
detected issues.}

\item{full_report}{Logical; if \code{TRUE}, generates a full report of the
validation process.}

\item{f_proof}{Logical; if \code{TRUE}, generates proof plots for data validation.}

\item{printPDF}{Logical; if \code{TRUE} and \code{f_proof} is \code{TRUE}, saves the plots
to a PDF file. If so, provide the desired path to output
the PDF file in the argument \code{out_qc_folder}.}

\item{out_qc_folder}{Optional; a string specifying the path to the folder
where output PDF should be saved (only if \code{printPDF = TRUE}).
Default: current working directory.}

\item{dmaqc_shipping_info}{(character) File path to the DMAQC file.
Only the BIC can use this argument.}

\item{dmaqc_phase2validate}{(character) Provide phase to validate. This argument
is not required since it should be extracted from the input folder or from the
new required file \code{metadata_phase.txt}. Please ignore. However, if this argument is provided,
it will take priority and this will be the phase.}

\item{validate_uniprot}{Logical; if \code{TRUE}, validates against the UniProt
database.}

\item{verbose}{Logical; if \code{TRUE}, prints detailed messages during validation.}
}
\value{
Depending on the settings, this function may return the number of
issues found, generate reports or plots, or simply perform the
validation without returning anything.
}
\description{
This function validates LAB assay data based on a set of criteria, including
folder structure, metadata, and file contents. It supports optional
functionalities like creating a PDF report and DMAQC validation
(data only available at the BIC).
}
\examples{
\dontrun{
validate_lab("/path/to/results", cas = "broad_rg", return_n_issues = TRUE)
}
}
