% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_lc_column_id}
\alias{validate_lc_column_id}
\title{Validate 'lc_column_id' column}
\usage{
validate_lc_column_id(df, column_name, verbose = TRUE)
}
\arguments{
\item{df}{A data frame that should contain the 'lc_column_id' column.}

\item{verbose}{A logical indicating whether to print informative messages.
Default is TRUE.}
}
\value{
An invisible NULL. The function is used mainly for its side effects
(i.e., printing validation results).
}
\description{
This function checks the 'lc_column_id' column of a provided data frame
to ensure that it exists, contains no NA values, and contains no spaces
in its entries. It also reports the number of unique values in the column.
}
\examples{
df <- data.frame(lc_column_id = c("id1", "id2", "id3", "id1", "id 2", NA))
validate_lc_column_id(df)

}
