% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_phase}
\alias{validate_phase}
\title{extract PHASE from input folder path}
\usage{
validate_phase(input_results_folder)
}
\arguments{
\item{input_results_folder}{(char) input_results_folder path}
}
\value{
(vector) PHASE code
}
\description{
extract ASSAY from input folder path
}
