% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_processFolder}
\alias{validate_processFolder}
\title{extract PROCESSED_YYYYMMDD folder from input folder path}
\usage{
validate_processFolder(input_results_folder)
}
\arguments{
\item{input_results_folder}{(char) input_results_folder path}
}
\value{
(vector) PROCESSED_YYYYMMDD folder name
}
\description{
extract PROCESSED_YYYYMMDD folder from input folder path
}
