% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_qc.R
\name{validate_proteomics}
\alias{validate_proteomics}
\title{Validate a Proteomics submissions}
\usage{
validate_proteomics(
  input_results_folder,
  isPTM,
  cas,
  dmaqc_shipping_info = NULL,
  f_proof = FALSE,
  out_qc_folder = NULL,
  return_n_issues = TRUE,
  full_report = FALSE,
  printPDF = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{input_results_folder}{(char) path to the PROCESSED folder to check}

\item{isPTM}{(logical) \code{TRUE} if it is Post-Translational-Modification proteomics assay}

\item{cas}{(char) CAS code}

\item{dmaqc_shipping_info}{(char) phase code}

\item{f_proof}{(char) print out pdf with charts including:
\itemize{
\item Reported Ion Intensity boxplot distribution and percentage of NA values per sample
\item Ratio: ratio boxplot distribution and percentage of NA values per samples
}}

\item{out_qc_folder}{(char) if f_proof is TRUE, then a folder must be provided}

\item{return_n_issues}{(logical) if \code{TRUE} (default) returns the number of issues}

\item{full_report}{(logical) if \code{FALSE} (default) it returns only the
total number of issues. If \code{TRUE} returns the details of the number of issues (by
group of files, e.g., results, metadata_metabolites, etc)}

\item{printPDF}{(logical) if \code{TRUE} (default print plots to pdf)}

\item{verbose}{(logical) \code{TRUE} (default) prints QC details.}
}
\value{
(data.frame) Summary of issues
}
\description{
Validate a Proteomics submission
}
