% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_tissue}
\alias{validate_tissue}
\title{extract and validate TISSUE CODE from input folder path}
\usage{
validate_tissue(input_results_folder)
}
\arguments{
\item{input_results_folder}{(char) input_results_folder path}
}
\value{
(vector) PHASE code
}
\description{
extract and validate TISSUE CODE from input folder path
}
