% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_two_phases}
\alias{validate_two_phases}
\title{validate a phase with two phases (pass1a and 1c)}
\usage{
validate_two_phases(phase_details, verbose = TRUE)
}
\arguments{
\item{phase_details}{(char) expected output of \code{set_phase}}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(char) the expected phase_details function
}
\description{
This function only works to validate two phases reported
as for example, 'PASS1A-06|PASS1C-06' using the separator '|'
}
