% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_yyyymmdd_dates}
\alias{validate_yyyymmdd_dates}
\title{Validate YYYY-MM-DD Dates in a Data Frame}
\usage{
validate_yyyymmdd_dates(df, date_column, verbose = TRUE)
}
\arguments{
\item{df}{A data frame that contains the date information to be validated.}

\item{date_column}{A character string specifying the name of the column in \code{df} that contains the dates to be validated.}

\item{verbose}{A logical value indicating whether or not to print messages (default: \code{TRUE}).}
}
\value{
number of issues found
}
\description{
Validate Dates in a Specified Column of a Data Frame
}
\details{
This function checks for the validity of dates in a specified column of a given data frame.
Valid dates are in the format YYYY-MM-DD, with year values between 2018 and 2026,
month values between 1 and 12, and day values between 1 and 31.
The function prints a list of invalid dates and a success message if all dates are valid.
}
\examples{
df <- data.frame(
  extraction_date = c("2022-01-31", "2023-12-01", "2027-11-30"),
  other_column = 1:3
)
validate_yyyymmdd_dates(df, "extraction_date")
}
