% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imm_qc.R
\name{write_imm_releases}
\alias{write_imm_releases}
\title{Write IMM Data Release}
\usage{
write_imm_releases(
  input_results_folder,
  folder_name = "motrpac_release",
  folder_root = NULL,
  version_file = "v1.0",
  verbose = TRUE
)
}
\arguments{
\item{input_results_folder}{(character) Path to the \code{PROCESSED_YYYYMMDD} folder
containing the IMM data.}

\item{folder_name}{(character) Output folder name. Default is \code{"motrpac_release"}.}

\item{folder_root}{(character) Absolute path where the output folder will be created.
Default is the current working directory.}

\item{version_file}{(character) File version number (e.g., \code{"v1.0"}). Default is \code{"v1.0"}.}

\item{verbose}{(logical) If \code{TRUE} (default), displays messages during processing.}
}
\value{
Creates the BIC release folder and file structure:
\verb{PHASE/OMICS/TCODE_NAME/ASSAY/}, and writes out files with names:
\code{motrpac_phase-code_tissuecode_assay_file-details-version.txt},
where \code{file-details} can be:
\itemize{
\item \code{metadata-analyte},
\item \code{metadata-samples},
\item \code{results}
}
}
\description{
Write out IMM data releases. This function doesn't check whether
data has been submitted according to guidelines; it assumes that the data
has already been validated and is ready for release.
}
\examples{
\dontrun{
write_imm_releases(
   input_results_folder = "/full/path/to/PROCESSED_YYYYMMDD/")
}
}
