% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_qc.R
\name{write_olink_releases}
\alias{write_olink_releases}
\title{Write olink data release}
\usage{
write_olink_releases(
  input_results_folder,
  folder_name = "motrpac_release",
  folder_root = NULL,
  version_file = "v1.0",
  verbose = TRUE
)
}
\arguments{
\item{input_results_folder}{(char) Path to the RESULTS_YYYYMMDD folder}

\item{folder_name}{(char) output folder name.}

\item{folder_root}{(char) absolute path to write the output folder.
Default: current directory}

\item{version_file}{(char) file version number (v#.#)}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
bic release folder/file structure
\verb{PHASE/OMICS/TCODE_NAME/ASSAY/} and file names, including:
\itemize{
\item \code{motrpac_phase-code_tissuecode_assay_file-details-version.txt}
where files-details can be:
\item \code{metadata-proteins},
\item \code{metadata-samples},
\item \code{results}
}
}
\description{
Write out olink data releases. Doesn't check whether
data has been submited according to guidelines
}
\examples{
\dontrun{
write_olink_releases(
   input_results_folder = "/full/path/to/RESULTS_YYYYMMDD/")
}
}
