% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assay_codes}
\alias{assay_codes}
\title{Motrpac assay codes}
\format{A data frame with assay codes
\describe{
\item{omics}{Main OMIC group}
\item{submission_code}{Assay submission code}
\item{assay_code}{Assay release code}
\item{assay}{Assay long name}
\item{site}{MoTrPAC chemical analysis site}
}}
\usage{
assay_codes
}
\description{
Motrpac assay codes
}
\examples{
\dontrun{
 assay_codes
}
}
\keyword{datasets}
