% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bic_animal_tissue_code}
\alias{bic_animal_tissue_code}
\title{BIC Tissue Code and Name}
\format{A data frame with tissue codes
\describe{
\item{bic_tissue_code}{\code{T} + integer, as encoded in vial_label by DMAQC}
\item{bic_tissue_name}{Tissue name (as provided by DMAQC)}
\item{motrpac_tissue_code}{Both code and tissue name}
\item{tissue_name_release}{Code + tissue name as used in releases}
\item{abbreviation}{Tissue abbreviation}
\item{tissue_hex_colour}{Tissue Hex Colour code}

}}
\usage{
bic_animal_tissue_code
}
\description{
BIC Tissue Code and Name
}
\examples{
\dontrun{
 bic_animal_tissue_code
}
}
\keyword{datasets}
