% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bic_animal_tissue_code}
\alias{bic_animal_tissue_code}
\title{BIC Tissue Code and Name}
\format{Data frame
\describe{
BIC code for aninal tissues
}}
\usage{
bic_animal_tissue_code
}
\description{
ssue codes and names generated by the BIC
}
\keyword{datasets}
