% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{check_manifest_rawdata}
\alias{check_manifest_rawdata}
\title{check rawfiles between manifest and metabolite_sample matches}
\usage{
check_manifest_rawdata(input_results_folder, m_s_n_raw, verbose = TRUE)
}
\arguments{
\item{input_results_folder}{(char) input path folder}

\item{m_s_n_raw}{(list) list of raw files available in the metadata sample file}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check rawfiles between manifest and metabolite_sample matches
}
