% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{check_metadata_metabolites}
\alias{check_metadata_metabolites}
\title{check metadata metabolites}
\usage{
check_metadata_metabolites(df, nameun, verbose = TRUE)
}
\arguments{
\item{df}{(data.frame) metadata_metabolites}

\item{nameun}{(char) specify whether \code{named} or \code{unnamed} files}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether metadata_metabolites is following guidelines
}
