% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{check_metadata_samples}
\alias{check_metadata_samples}
\title{check metadata samples}
\usage{
check_metadata_samples(df, cas, return_n_issues = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{(data.frame) metadata_metabolites}

\item{cas}{(char) CAS site code}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues.}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) number of issues identified
}
\description{
check whether metadata_sample is following guidelines
}
