% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dl_read_gcp}
\alias{dl_read_gcp}
\title{dl_read_gcp: Data Load, Read file from Google Cloud}
\usage{
dl_read_gcp(
  path,
  sep = "\\t",
  header = TRUE,
  tmpdir,
  gsutil_path = "gsutil",
  check_first = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{path}{(char) GCS path, i.e., starts with "gs://"}

\item{sep}{(char) column separator to use with \link[data.table:fread]{data.table::fread}}

\item{header}{(bool) whether input file has a header line}

\item{tmpdir}{(char) scratch directory to download files from GCS}

\item{gsutil_path}{(char) path to \code{gsutil} on your computer.
Can be "gsutil" if \code{gsutil} is in your \code{$PATH}.}

\item{check_first}{(char) check if file exists in \code{tmpdir} before
downloading it. Read in existing file if it exists.
Should be set to \code{TRUE} if you are running this function in parallel.}

\item{verbose}{(logical) \code{TRUE} shows messages (default \code{FALSE})}

\item{...}{optional arguments for \link[data.table:fread]{data.table::fread}}
}
\value{
a data table
}
\description{
Read a single file from Google Cloud Storage (GSC) into a data table
}
\examples{
\dontrun{
pheno = dl_read_gcp(path = "gs://your-bucket/file.txt")
}
}
