% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{generate_phase_details}
\alias{generate_phase_details}
\title{Generate the phase detail for submissions}
\usage{
generate_phase_details(phase_metadata, verbose = TRUE)
}
\arguments{
\item{phase_metadata}{(char) expected output of \code{set_phase}}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(char) the expected phase_details function
}
\description{
The phase details is as simple as creating a lower case version
of the phase. However, in case of PASS1A/1C a new version has to be generated:
pass1ac-06
This function detects whether there are two phases, and if so,
generate the expected version: either pass1ac-06 or pass1ac-18
}
