% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_data_dictionary.R
\name{get_and_validate_mdd}
\alias{get_and_validate_mdd}
\title{Get and validate Metabolomics Data Dictionary from Metabolomics Workbench}
\usage{
get_and_validate_mdd(remove_duplications = FALSE)
}
\arguments{
\item{remove_duplications}{(logical) if `TRUE``, removes duplications.}
}
\value{
(vector) PHASE code
}
\description{
Get and validate Metabolomics Data Dictionary from
Metabolomics Workbench
}
