% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{set_phase}
\alias{set_phase}
\title{Set the phase to be validated.}
\usage{
set_phase(input_results_folder, dmaqc_phase2validate, verbose = TRUE)
}
\arguments{
\item{input_results_folder}{(char) path to the PROCESSED/RESULTS folder to check}

\item{dmaqc_phase2validate}{(data.frame) dmaqc shipping information}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(int) the phase to be validated.
}
\description{
A group might choose to combine two different phases, due to
the complications associated with PASS1A/1C. If they choose to combine
two phases, the CAS must provide a new file \code{metadata_phase.txt} with a single
line, as for example: \code{PASS1A-06|PASS1C-06}. This function checks if the
file is available, and set that phase as the phases to validate. In summary,
the order of preference is:
\enumerate{
\item function's argument: dmaqc_phase2validate (if provided in the validation functions)
\item \code{metadata_phase.txt} file if available in the batch folder.
\item Phase in folder structure
}
}
