% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_batch}
\alias{validate_batch}
\title{extract BATCH_YYYYMMDD folder}
\usage{
validate_batch(input_results_folder)
}
\arguments{
\item{input_results_folder}{(char) input_results_folder path}
}
\value{
(vector) BATCH_YYYYMMDD folder name
}
\description{
extract BATCH_YYYYMMDD folder from input folder path
}
