% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{validate_metabolomics}
\alias{validate_metabolomics}
\title{Validate a Metabolomics submission}
\usage{
validate_metabolomics(
  input_results_folder,
  cas,
  tissue_code,
  phase,
  dmaqc_shipping_info,
  verbose = TRUE
)
}
\arguments{
\item{input_results_folder}{(char) path to the PROCESSED folder to check}

\item{cas}{(char) CAS code}

\item{tissue_code}{(char) tissue code}

\item{phase}{(char) phase code}

\item{dmaqc_shipping_info}{(char) phase code}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(data.frame) Summary of issues
}
\description{
Validate a Metabolomics submission
}
