% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{validate_metabolomics}
\alias{validate_metabolomics}
\title{Validate a Metabolomics submission}
\usage{
validate_metabolomics(
  input_results_folder,
  cas,
  dmaqc_shipping_info = NULL,
  return_n_issues = FALSE,
  full_report = FALSE,
  verbose = TRUE,
  f_proof = FALSE,
  out_qc_folder = NULL,
  printPDF = TRUE
)
}
\arguments{
\item{input_results_folder}{(char) path to the PROCESSED folder to check}

\item{cas}{(char) CAS code}

\item{dmaqc_shipping_info}{(char) phase code}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues}

\item{full_report}{(logical) if \code{FALSE} (default) it returns only the
total number of issues. If \code{TRUE} returns the details of the number of issues (by
group of files, e.g., results, metadata_metabolites, etc)}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}

\item{f_proof}{(char) print out pdf with charts including:
\itemize{
\item Sample intensity distribution
\item Unique ID counts
\item NA values
}}

\item{out_qc_folder}{(char) output qc folder (it creates the folder if it doesn't exist)}

\item{printPDF}{(logical) if \code{TRUE} (default print plots to pdf)}
}
\value{
(data.frame) Summary of issues
}
\description{
Validate a Metabolomics submission
}
