% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{validate_metabolomics}
\alias{validate_metabolomics}
\title{Validate a Metabolomics submission}
\usage{
validate_metabolomics(
  input_results_folder,
  cas,
  return_n_issues = FALSE,
  full_report = FALSE,
  f_proof = FALSE,
  printPDF = TRUE,
  out_qc_folder = NULL,
  dmaqc_shipping_info = NULL,
  dmaqc_phase2validate = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input_results_folder}{(char) path to the PROCESSED folder to check}

\item{cas}{(char) CAS code}

\item{return_n_issues}{(logical) if \code{TRUE} returns the number of issues}

\item{full_report}{(logical) if \code{FALSE} (default) it returns only the
total number of issues. If \code{TRUE} returns the details of the number of issues (by
group of files, e.g., results, metadata_metabolites, etc)}

\item{f_proof}{(logical) generate charts including:
\itemize{
\item Sample intensity distribution
\item Unique ID counts
\item NA values
}}

\item{printPDF}{(logical) if \code{f_proof = TRUE}, then \code{ printPDF = TRUE} (default)
prints plots to pdf file and argument \code{out_qc_folder} should be provided.}

\item{out_qc_folder}{(char) Folder to save the pdfs (if \code{printPDF = TRUE}).
It will create the folder if it doesn't exist. If this argument is not provided,
the files will be written to the working directory}

\item{dmaqc_shipping_info}{(char) File path to the DMAQC file. Only the BIC can use this argument}

\item{dmaqc_phase2validate}{(char) Provide phase to validate. This argument
is not required since it should be extracted from the input folder or from the
new required file \code{metadata_phase.txt}. Please, ignore.
However, if this argument is provided,
it will take priority and this will be the phase.
\code{metadata_phase.txt} will be ignored). Examples
\itemize{
\item Folder with \code{PASS1A-06}: type either \code{PASS1A-06} or leave it \code{NULL}
\item Both \code{PASS1A-06} and \code{PASS1C-06}: type \code{PASS1A-06|PASS1C-06}
\item Only \code{PASS1C-06}: type \code{PASS1C-06}
}}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
(data.frame) Summary of issues
}
\description{
Validate a Metabolomics submission
}
