% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_data_dictionary.R
\name{validate_refmetname}
\alias{validate_refmetname}
\title{Validate refmet_name}
\usage{
validate_refmetname(dataf, verbose)
}
\arguments{
\item{dataf}{(data.frame) Data frame with the refmet_name column}

\item{verbose}{(logical) \code{TRUE} (default) shows messages)}
}
\value{
(numeric) number of refmet_name ids not available in RefMet
}
\description{
Validate the refment_name using the Metabolomics Workbench API
}
