% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_qc.R
\name{write_metabolomics_releases}
\alias{write_metabolomics_releases}
\title{Write metabolomics data release}
\usage{
write_metabolomics_releases(
  input_results_folder,
  cas,
  folder_name = "motrpac_release",
  folder_root = NULL,
  version_file = "v1.0",
  verbose = TRUE
)
}
\arguments{
\item{input_results_folder}{(char) Path to the PROCESSED_YYYYMMDD folder}

\item{cas}{(char) Chemical Analytical Site code (e.g "umichigan")}

\item{folder_name}{(char) output files name. Must have a \code{.yaml} extension.}

\item{folder_root}{(char) absolute path to write the output files.
Default: current directory}

\item{version_file}{(char) file version number (\code{v#.#})}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
bic release folder/file structure
\verb{PHASE/OMICS/TCODE_NAME/ASSAY/} and file names, including:
\itemize{
\item \code{motrpac_YYYYMMDD_phasecode_tissuecode_omics_assay_file-details.txt}
where files-details can be:
\item \code{named-experimentalDetails.txt}
\item \code{named-metadata-metabolites.txt}
\item \code{metadata-samples.txt}
\item \code{named-results.txt}
}
}
\description{
Write out metabolomics data releases. Doesn't check whether
data has been submited according to guidelines
}
\examples{
\dontrun{
write_metabolomics_releases(input_results_folder = "/path/to/PROCESSED_YYYYMMDD/", 
cas = "umichigan")
}
}
