% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_qc.R
\name{write_proteomics_releases}
\alias{write_proteomics_releases}
\title{Write proteomics data release}
\usage{
write_proteomics_releases(
  input_results_folder,
  folder_name = "motrpac_release",
  folder_root = NULL,
  version_file = "v1.0",
  verbose = TRUE
)
}
\arguments{
\item{input_results_folder}{(char) Path to the PROCESSED_YYYYMMDD folder}

\item{folder_name}{(char) output folder name.}

\item{folder_root}{(char) absolute path to write the output folder. Default: current directory}

\item{version_file}{(char) file version number (v#.#)}

\item{verbose}{(logical) \code{TRUE} (default) shows messages}
}
\value{
bic release folder/file structure \verb{PHASE/OMICS/TCODE_NAME/ASSAY/} and file names, including:
\code{motrpac_YYYYMMDD_phasecode_tissuecode_omics_assay_file-details.txt} where files-details can be:
\code{rii-results.txt}, \code{ratio-results.txt}, \code{vial-metadata.txt}
}
\description{
Write out proteomics data releases. Doesn't check whether
data has been submited according to guidelines
}
\examples{
\dontrun{
write_proteomics_releases(
   input_results_folder = "/full/path/to/RESULTS_YYYYMMDD/")
}
}
